refactor(popup): Move v1/v2 routing to PopupLayer
Move the isThreadedAnnotation branching from PopupReply to PopupLayer
so PopupReply is a clean v1-only component. PopupLayer now renders
PopupReplyV2 or PopupReply directly based on the feature flag.
