+++ src/components/Popups/PopupReply.tsx
@@ -3,39 +3,25 @@ import Popper from '@popperjs/core';
 import { useSelector } from 'react-redux';
 import { useIntl } from 'react-intl';
 import FocusTrap from 'box-ui-elements/es/components/focus-trap/FocusTrap';
+
+import ReplyForm from '../ReplyForm';
+
+import { getRotation, getScale } from '../../store/options';
+
 import messages from './messages';
 import PopupBase from './PopupBase';
-import PopupReplyV2 from './PopupReplyV2';
-import ReplyForm from '../ReplyForm';
-import { getScale, getRotation } from '../../store/options';
 import { PopupReference } from './Popper';
 import './PopupReply.scss';
 
-type BaseProps = {
+export type Props = {
     className?: string;
-};
-
-type LegacyProps = BaseProps & {
     isPending: boolean;
-    isThreadedAnnotation?: false;
     onCancel: (text?: string) => void;
     onChange: (text?: string) => void;
     onSubmit: (text: string) => void;
     reference: PopupReference;
     value?: string;
 };
-
-type ThreadedProps = BaseProps & {
-    isPending: boolean;
-    isThreadedAnnotation: true;
-    onCancel: (text?: string) => void;
-    onChange: (text?: string) => void;
-    onSubmit: (text: string) => void;
-    reference: PopupReference;
-    value?: string;
-};
-
-export type Props = LegacyProps | ThreadedProps;
 
 const isIE = (): boolean => {
     const { userAgent } = navigator;
@@ -97,7 +83,7 @@ const getOptions = (): Partial<Popper.Options> => {
 export default function PopupReply(props: Props): JSX.Element {
     const intl = useIntl();
     const popupRef = React.useRef<PopupBase>(null);
-    const popupOptions = React.useRef<Partial<Popper.Options>>(getOptions()); // Keep the options reference the same between renders
+    const popupOptions = React.useRef<Partial<Popper.Options>>(getOptions());
     const rotation = useSelector(getRotation);
     const scale = useSelector(getScale);
 
@@ -109,17 +95,7 @@ export default function PopupReply(props: Props): JSX.Element {
         }
     }, [popupRef, rotation, scale]);
 
-    if (props.isThreadedAnnotation) {
-        const { onSubmit, reference } = props as ThreadedProps;
-        return (
-            <PopupReplyV2
-                onSubmit={onSubmit}
-                reference={reference}
-            />
-        );
-    }
-
-    const { isPending, onCancel, onChange, onSubmit, value = '', isThreadedAnnotation: _, ...rest } = props as LegacyProps;
+    const { isPending, onCancel, onChange, onSubmit, value = '', ...rest } = props;
 
     return (
         <FocusTrap>
